+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Method `firstDayOfFirstWeek()` - returns start of the first day of the first week in year.
 * Method `getWeekNumber()` - returns number of the week in year.
 * Method `getDaysDifference()` - returns count of days between two dates.
+* Method `getLastWeekNumber()` - returns number of the last week in year.
 
 ## [0.1.2] - 2020-11-02